Move microsyntaxes out of alfa-parser
