+++ packages/alfa-parser/src/index.ts
@@ -1,2 +1,1 @@
-export * from "./microsyntaxes";
 export * from "./parser";
