+++ packages/alfa-table/src/helpers.ts
@@ -1,7 +1,8 @@
 import { Element, Attribute } from "@siteimprove/alfa-dom";
 import { clamp } from "@siteimprove/alfa-math";
-import { parseNonNegativeInteger } from "@siteimprove/alfa-parser";
 import { Ok } from "@siteimprove/alfa-result";
+
+import { parseNonNegativeInteger } from "./microsyntaxes";
 
 const { parseAttribute } = Attribute;
